mv scaling defs to separate header
